backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.9
pydantic>=2.9.0
bcrypt>=4.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
